Fixed rgb colours being wrong.

+++ Program/Converter.h
@@ -10,7 +10,7 @@
 class Converter{
 public:
     static std::string rgb2hex(int r, int g, int b, bool with_head);
-    static int* hex2rgb(std::string hex);
+    static double* hex2rgb(std::string hex);
     static int hex2int(std::string hex);
     static std::string int2hex(int);
     static std::string char2hex(char* bytes, size_t size);
